requirements.txt
astroid==2.0.4
bcrypt==3.1.4
blinker==1.4
cffi==1.11.5
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
Pillow==5.3.0
pycparser==2.19
pylint==2.1.1
six==1.11.0
SQLAlchemy==1.2.14
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
